pyproject.toml
[project]
name = "sailowtech-ctd"
version = "0.1.2"
description = ""
authors = [
    { name = "Arthur Jacobs", email ="arthur.jacobs@sailowtech.ch" },
    { name = "Mauro Sbicego" }
]
readme = "README.md"

requires-python = "=3.12.8"
dependencies = [
"pyyaml>=6.0.2",
"smbus2>=0.4.3",
"fastapi[standard] (>=0.115.8,<0.116.0)",
"uvicorn[standard] (>=0.34.0,<0.35.0)",
"atlas-i2c (>=0.3.1,<0.4.0)",
"peewee (>=3.17.9,<4.0.0)"
]

[project.scripts]
web = 'software.sailowtech_ctd.__main__:start'
web-debug = 'software.sailowtech_ctd.__main__:start_debug'

[project.urls]
"Documentation" = "https://sailowtech.github.io/Sailowtech-CTD/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "software/sailowtech_ctd" }]
